requirements.txt
bottle==0.12.8
docker-py==1.1.0
PyYAML==3.11
